Use js-search as search engine (replace fuse)
