+++ package.json
@@ -2,11 +2,7 @@
   "name": "laa-apply-for-legal-aid",
   "private": true,
   "dependencies": {
-    "@rails/webpacker": "3.5",
-    "fuse.js": "^3.3.0",
-    "govuk-frontend": "^1.1.1"
-  },
-  "devDependencies": {
-    "webpack-dev-server": "^3.1.10"
+    "govuk-frontend": "^1.1.1",
+    "js-search": "^1.4.2"
   }
 }
